Clanup same-package imports
