+++ packages/ui-app/src/AddressMini.tsx
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { BareProps } from '@polkadot/ui-app/types';
+import { BareProps } from './types';
 
 import BN from 'bn.js';
 import React from 'react';
