+++ packages/ui-app/src/Button/Group.tsx
@@ -7,11 +7,9 @@ import { GroupProps, GroupType } from './types';
 import React from 'react';
 import SUIButton from 'semantic-ui-react/dist/commonjs/elements/Button/Button';
 
-import Button from '@polkadot/ui-app/Button';
 import classes from '../util/classes';
 import Divider from './Divider';
 
-// ({ children, className, style }: GroupProps)
 class ButtonGroup extends React.PureComponent<GroupProps> {
   render () {
     const { children, className, style } = this.props;
@@ -22,7 +20,7 @@ class ButtonGroup extends React.PureComponent<GroupProps> {
         style={style}
       >
         <SUIButton.Group>
-          <Button.Group.Divider style={{ padding: '0em' }} />
+          <Divider style={{ padding: '0em' }} />
           {children}
         </SUIButton.Group>
       </div>
