+++ packages/ui-app/src/InputFile.tsx
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { WithNamespaces } from 'react-i18next';
-import { BareProps } from '@polkadot/ui-app/types';
+import { BareProps } from './types';
 
 import React from 'react';
 import Dropzone from 'react-dropzone';
