+++ packages/ui-app/src/Params/Param/Vector.tsx
@@ -7,8 +7,8 @@ import { TypeDef } from '@polkadot/types/codec';
 import { Props as BareProps, RawParam } from '../types';
 
 import React from 'react';
-import { Button } from '@polkadot/ui-app/index';
 
+import Button from '../../Button';
 import translate from '../../translate';
 import getInitValue from '../initValue';
 import Bare from './Bare';
