+++ packages/ui-app/src/Status/index.tsx
@@ -2,14 +2,15 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { I18nProps } from '@polkadot/ui-app/types';
+import { I18nProps } from '../types';
 import { QueueStatus, QueueTx, QueueTx$Status } from './types';
 
 import React from 'react';
-import { AddressMini, Icon } from '@polkadot/ui-app/index';
-import classes from '@polkadot/ui-app/util/classes';
 import { Method } from '@polkadot/types';
 
+import AddressMini from '../AddressMini';
+import Icon from '../Icon';
+import classes from '../util/classes';
 import translate from '../translate';
 
 type Props = I18nProps & {
